+++ docs/changelog.rst
@@ -130,6 +130,9 @@ Tracklist controller
   :meth:`mopidy.core.TracklistController.filter`.
   Use the ``criteria`` argument instead.
 
+- Removed the support for passing filter criterias as keyword arguments to
+  :meth:`mopidy.core.TracklistController.remove`.
+  Use the ``criteria`` argument instead.
 
 Backend API
 -----------